requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
fonttools==4.51.0
greenlet==3.0.3
htmlmin==0.1.12
idna==3.7
ImageHash==4.3.1
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
multimethod==1.11.2
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
phik==0.12.4
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.6.0
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
requests==2.31.0
scipy==1.11.4
seaborn==0.12.2
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.29
stack-data==0.6.3
statsmodels==0.14.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
visions==0.7.6
wcwidth==0.2.13
widgetsnbextension==4.0.10
wordcloud==1.9.3
ydata-profiling==4.7.0
